[MIG] product_harmonized_system to v16
