+++ product_harmonized_system/i18n/product_harmonized_system.pot
@@ -4,7 +4,7 @@
 #
 msgid ""
 msgstr ""
-"Project-Id-Version: Odoo Server 15.0\n"
+"Project-Id-Version: Odoo Server 16.0\n"
 "Report-Msgid-Bugs-To: \n"
 "Last-Translator: \n"
 "Language-Team: \n"
@@ -49,20 +49,6 @@ msgstr ""
 #. module: product_harmonized_system
 #: model:ir.model.fields,field_description:product_harmonized_system.field_hs_code__company_id
 msgid "Company"
-msgstr ""
-
-#. module: product_harmonized_system
-#: model:ir.model.fields,field_description:product_harmonized_system.field_product_product__origin_state_id
-#: model:ir.model.fields,field_description:product_harmonized_system.field_product_template__origin_state_id
-msgid "Country State of Origin"
-msgstr ""
-
-#. module: product_harmonized_system
-#: model:ir.model.fields,help:product_harmonized_system.field_product_product__origin_state_id
-#: model:ir.model.fields,help:product_harmonized_system.field_product_template__origin_state_id
-msgid ""
-"Country State of origin of the product.\n"
-"This field is used for the Intrastat declaration, selecting one of the Northern Ireland counties will set the code 'XI' for products from the United Kingdom whereas code 'XU' will be used for the other UK counties."
 msgstr ""
 
 #. module: product_harmonized_system
@@ -171,7 +157,7 @@ msgid "Printed circuits"
 msgstr ""
 
 #. module: product_harmonized_system
-#: model:ir.model,name:product_harmonized_system.model_product_product
+#: model:ir.model,name:product_harmonized_system.model_product_template
 msgid "Product"
 msgstr ""
 
@@ -192,13 +178,13 @@ msgid "Product Category"
 msgstr ""
 
 #. module: product_harmonized_system
-#: model:ir.model,name:product_harmonized_system.model_product_template
-msgid "Product Template"
+#: model:ir.model.fields,field_description:product_harmonized_system.field_hs_code__product_tmpl_count
+msgid "Product Tmpl Count"
 msgstr ""
 
 #. module: product_harmonized_system
-#: model:ir.model.fields,field_description:product_harmonized_system.field_hs_code__product_tmpl_count
-msgid "Product Tmpl Count"
+#: model:ir.model,name:product_harmonized_system.model_product_product
+msgid "Product Variant"
 msgstr ""
 
 #. module: product_harmonized_system
