+++ product_harmonized_system/models/__init__.py
@@ -1,3 +1,4 @@
 from . import hs_code
 from . import product_category
 from . import product_template
+from . import product_product
