+++ product_harmonized_system/views/product_category.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8" ?>
 <!--
-  Copyright 2010-2020 Akretion (http://www.akretion.com/)
+  Copyright 2010-2022 Akretion (http://www.akretion.com/)
   @author Alexis de Lattre <alexis.delattre@akretion.com>
   License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 -->
@@ -11,11 +11,13 @@
         <field name="model">product.category</field>
         <field name="inherit_id" ref="product.product_category_form_view" />
         <field name="arch" type="xml">
-            <xpath expr="//group[@name='first']" position="after">
-                <group name="hs_code" string="Import/Export Properties">
-                    <field name="hs_code_id" />
+            <group name="first" position="after">
+                <group name="hs_code">
+                    <group name="hs_code_l2" string="Import/Export Properties">
+                        <field name="hs_code_id" />
+                    </group>
                 </group>
-            </xpath>
+            </group>
         </field>
     </record>
 </odoo>
